requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==23.1.0
automium==0.2.6
automium-web==0.1.1
backports.entry-points-selectable==1.1.0
BareNecessities==0.2.8
bcrypt==4.0.1
beautifulsoup4==4.11.2
blog==1.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.6
CommandTool==0.5.3
contact==1.0.5
distlib==0.3.3
Django==4.0.6
django-admin-volt==1.0.10
django-ckeditor==6.5.1
django-froala-editor==4.0.13
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-material-admin==1.8.6
django-richtextfield==1.6.1
django-social-share==2.3.0
django-tinymce==3.5.0
django-verified-email-field==1.9.0
djangorestframework==3.13.1
dnspython==2.3.0
docopt==0.6.2
environment==1.0.0
exceptiongroup==1.1.1
filelock==3.3.0
Flask==2.2.3
fontawesomefree==6.3.0
Gallery==0.1.0
grid==0.7.1
h11==0.14.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.2
outcome==1.2.0
Pillow==9.2.0
pipreqs==0.4.13
platformdirs==2.4.0
psycopg2==2.9.6
pycparser==2.21
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.14.0
pymdown-extensions==9.10b3
pymongo==4.3.3
pyotp==2.8.0
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.1
urllib3==1.26.7
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
Werkzeug==2.2.3
wsproto==1.2.0
yarg==0.1.9
